% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEKS_decomp.R
\name{get_single_weight}
\alias{get_single_weight}
\title{Get weight between times t and period of interest}
\usage{
get_single_weight(
  data_window_a,
  data_window_b,
  all_id,
  id_of_interest,
  period_a,
  period_b
)
}
\arguments{
\item{data_window_a}{vector of the times at which price observations were made. \cr
NOTE: \code{times} must be of class Date or numeric.}

\item{data_window_b}{vector of log of prices at the given time}

\item{all_id}{vector of distinct identification number of consumer goods}

\item{id_of_interest}{vector of expenditure weights used in the regressions}

\item{period_a}{either empty (will assume latest periods) or a vector of length
two relating to times compared where [1] is the 'start_time' time and [2] is the 'to' time
e.g Comparing contribution start_time "1973-04-01" to "1973-05-01" would be c("1973-04-01", "1973-05-01")}

\item{period_b}{optional. Single number for length of window for the data that
regressions are fit on. Note if window_length is present it assumes custom time is latest two periods
in window.}
}
\description{
Get weight between times t and period of interest
}
