% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEKS_decomp.R
\name{get_weight_dot}
\alias{get_weight_dot}
\title{Get weight between all times t and period of interest}
\usage{
get_weight_dot(data_window, id_of_interest, period)
}
\arguments{
\item{data_window}{vector of the times at which price observations were made. \cr
NOTE: \code{times} must be of class Date or numeric.}

\item{id_of_interest}{vector of log of prices at the given time}

\item{period}{vector of distinct identification number of consumer goods}
}
\description{
Get weight between all times t and period of interest
}
